requirements_v2.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch>=2.6.0
open_clip_torch>=2.20.0
multilingual-clip>=1.0.10
transformers>=4.21.0
pillow>=10.0.1
requests>=2.31.0
pydantic>=2.5.0